+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 ### New in 3.1.0 - (Released 2015/02/12)
-* Fixed signature of `Observable.create` to better fit F#.
+* Fixed signature of `Observable.create` to better fit F#. #56
+* Marked `ObservableBuilder.Return` and `ObservableBuilder.ReturnFrom` as `Obsolete` as users should use `Yield` and `YieldFrom`, which allow for multiple results. #53
 
 ### New in 3.0.1 - (Released 2014/12/07)
 * #48 Some of the Observable functions aren't very composable